requirements.txt
alembic==0.6.2
APScheduler==3.8.1
apturl==0.5.2
argcomplete==1.12.3
asgiref==3.5.2
attrs==22.2.0
awscli==1.27.6
awsebcli==3.20.3
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
binance-futures==1.1.0
blessed==1.19.1
blinker==1.3
botocore==1.23.54
Brlapi==0.7.0
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==7.1.2
cloudinary==1.26.0
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
dbfn==1.0.6
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.8
django-anymail==8.5
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap3==15.0.0
django-bootstrap4==21.2
django-bootstrap5==21.3
django-compressor==3.1
django-crispy-forms==1.13.0
django-daraja==1.2.0
django-extensions==3.1.5
django-filter==22.1
django-heroku==0.3.1
django-mpesa==2.0.10
django-paypal==1.1.2
django-registration-redux==2.9
django-reviews==1.2.1
django-sass-processor==1.1
django-static-precompiler==2.0
django-tinymce==3.3.0
djangorestframework==3.13.1
djangorestframework-jsonapi==6.0.0
dnspython==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
dominate==2.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.12.2
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.5.7
Flask-DebugToolbar==0.11.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Moment==0.5.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-Social==1.6.2
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.0
Flask-WTF==0.12
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflection==0.5.1
is-disposable-email==1.0.0
itsdangerous==0.23
Jinja2==2.7.3
jmespath==0.9.4
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libsass==0.21.0
lockfile==0.12.2
louis==3.12.0
lts==0.0.2
macaroonbakery==1.3.1
mailchimp-marketing==3.0.74
Mako==0.9.1
Markdown==3.3.6
MarkupSafe==0.18
monotonic==1.5
MouseInfo==0.1.3
Naked==0.1.31
numpy==1.23.2
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
pandas==1.4.4
paramiko==2.6.0
passlib==1.7.4
pathspec==0.9.0
pexpect==4.6.0
Pillow==9.0.0
pipenv==2023.7.11
platformdirs==3.9.1
plyer==2.0.0
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.4.2
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
pyjokes==0.6.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.6.8
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-decouple==3.5
python-dotenv==0.19.2
python3-xlib==0.15
pyttsx3==2.90
pytube==12.1.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pyxdg==0.26
PyYAML==5.3.1
rcssmin==1.1.0
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
rjsmin==1.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.6.0
SecretStorage==2.3.1
semantic-version==2.8.5
shellescape==3.8.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.1
SQLAlchemy==0.9.9
sqlparse==0.4.2
systemd-python==234
termcolor==1.1.0
texttable==1.6.7
trollius==2.1.post2
tzdata==2021.5
tzlocal==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
virtualenv==20.24.1
virtualenv-clone==0.5.7
visitor==0.1.3
wadllib==1.3.3
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==0.10.4
whitenoise==5.3.0
wikipedia==1.4.0
WTForms==1.0.5
wtforms-validators==1.0.0
xkit==0.0.0
zipp==3.7.0
